Improvements for CE lightmapping
Reverse order of options in CE slider as 0.000001 is actually the max and 0.999999 is the lowest.

Add labels for order of quality for CE lightmapping

Remove unused directives.

Make get_default_path args easier to understand.

Limit instance count to available logical processors on system.

Add property for checking if a build is a version of Halo 2.

Add a FilePicker.Option for JMS and ASS so that CE and H2 builds can have different descriptions in file dialog.

Remove whitespace throughout the project

Remove some useless settings.

Update some strings.