Fix a few more launcher issues
Edit the CancelableProgressBarWindow.xaml. No major visual changes really. Just editing the code a bit and hoping to solve a bug that happens randomly. Sometimes the task bar progress window will be stretched on one side with a black background. Can't reproduce this so taking the shotgun approach.

 Move owner property to the places the menu is opened from instead. Setting it from the window class was causing an issue where closing the task progress window would minimize the MainWindow

Disable the update button once it was pressed once.  If the Github releases list didn't respond fast enough then it was possible to click the button multiple times to get multiple dialog windows.

Set proper platform for Halo 2 cache packaging along with a comment saying the list of valid values.