Replace build_type with boolean.
In the olden days the launcher had to support many types of builds. These days only `MCC` and `Standalone` remain and someone already reused the MCC option for H3+ to flag whatever it's the newer version of that game generation.

Kept that behavior for now by converting this `enum` to a (nullable for technical reasons) boolean `alt_build`.